Split types out to separate files that need them
Now nothing but HAL types should be included in anything wpilibc
