+++ hal/src/main/native/include/hal/I2C.h
@@ -11,15 +11,13 @@
 
 #include "hal/Types.h"
 
+#include "hal/I2CTypes.h"
+
 /**
  * @defgroup hal_i2c I2C Functions
  * @ingroup hal_capi
  * @{
  */
-
-// clang-format off
-HAL_ENUM(HAL_I2CPort) { HAL_I2C_kInvalid = -1, HAL_I2C_kOnboard, HAL_I2C_kMXP };
-// clang-format on
 
 #ifdef __cplusplus
 extern "C" {
