+++ hal/src/main/native/include/hal/SPI.h
@@ -18,17 +18,6 @@
  * @{
  */
 
-// clang-format off
-HAL_ENUM(HAL_SPIPort) {
-  HAL_SPI_kInvalid = -1,
-  HAL_SPI_kOnboardCS0,
-  HAL_SPI_kOnboardCS1,
-  HAL_SPI_kOnboardCS2,
-  HAL_SPI_kOnboardCS3,
-  HAL_SPI_kMXP
-};
-// clang-format on
-
 #ifdef __cplusplus
 extern "C" {
 #endif
