+++ simulation/halsim_ds_socket/src/main/native/include/DSCommJoystickPacket.h
@@ -9,7 +9,7 @@
 
 #include <cstring>
 
-#include <hal/DriverStation.h>
+#include <hal/DriverStationTypes.h>
 
 namespace halsim {
