+++ wpilibc/src/main/native/include/frc/CAN.h
@@ -9,7 +9,7 @@
 
 #include <stdint.h>
 
-#include <hal/CANAPI.h>
+#include <hal/CANAPITypes.h>
 #include <wpi/ArrayRef.h>
 
 #include "frc/ErrorBase.h"
